Fix up selenium tests
Can now run them by running " pytest tests " in src
