+++ src/tests/test_selenium.py
@@ -11,6 +11,27 @@ from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 
 
+class TestLogin:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+
+    def teardown_method(self, method):
+        self.driver.quit()
+
+    def test_login(self):
+        self.driver.get("http://127.0.0.1:5000/auth/login")
+        self.driver.set_window_size(1245, 1040)
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+
+        self.driver.implicitly_wait(1)
+        
+        self.driver.find_element(By.CSS_SELECTOR, ".alert-success").click()
+
 class TestPost:
     def setup_method(self, method):
         self.driver = webdriver.Chrome()
@@ -28,12 +49,15 @@ class TestPost:
         actions.move_to_element(element).perform()
         element = self.driver.find_element(By.CSS_SELECTOR, "body")
         actions = ActionChains(self.driver)
-        actions.move_to_element(element, 0, 0).perform()
+        actions.move_to_element(element).perform()
         self.driver.find_element(By.ID, "login-email").click()
         self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
         self.driver.find_element(By.ID, "login-password").click()
         self.driver.find_element(By.ID, "login-password").send_keys("password123")
         self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+
+        self.driver.implicitly_wait(10)
+
         elements = self.driver.find_elements(By.CSS_SELECTOR, ".alert-success")
         assert len(elements) > 0
         self.driver.find_element(
@@ -44,6 +68,9 @@ class TestPost:
         self.driver.find_element(By.NAME, "content").click()
         self.driver.find_element(By.NAME, "content").send_keys("Testing 123")
         self.driver.find_element(By.CSS_SELECTOR, "input:nth-child(7)").click()
+
+        self.driver.implicitly_wait(10)
+        
         elements = self.driver.find_elements(By.CSS_SELECTOR, ".alert-success")
         assert len(elements) > 0
         elements = self.driver.find_elements(
